+++ sqlstore/src/main/resources/db/migration/V0940__auto_main.sql
@@ -1,3 +1,17 @@
+-- pool_order_changelog
+CREATE TABLE PoolOrderChangeLog (
+  poolOrderChangeLogId bigint(20) PRIMARY KEY AUTO_INCREMENT,
+  poolOrderId bigint(20) NOT NULL,
+  columnsChanged varchar(500) NOT NULL,
+  userId bigint(20) NOT NULL,
+  message longtext NOT NULL,
+  changeTime timestamp DEFAULT CURRENT_TIMESTAMP,
+  CONSTRAINT fk_poolOrderChangeLog_pool FOREIGN KEY (poolOrderId) REFERENCES PoolOrder(poolOrderId),
+  CONSTRAINT fk_poolOrderChangeLog_user FOREIGN KEY (userId) REFERENCES User(userId)
+) Engine=InnoDB DEFAULT CHARSET=utf8;
+
+
+-- pool_order_create_trigger
 -- StartNoTest
 DELIMITER //
 
@@ -43,3 +57,10 @@ FOR EACH ROW
 
 DELIMITER ;
 -- EndNoTest
+
+-- sequencing_parameters_paired
+ALTER TABLE SequencingParameters ADD COLUMN readLength2 int;
+UPDATE SequencingParameters SET readLength2 = paired * readLength;
+ALTER TABLE SequencingParameters MODIFY readLength2 int NOT NULL;
+ALTER TABLE SequencingParameters DROP COLUMN paired;
+