+++ sqlstore/src/main/resources/db/migration/V8888__sequencing_parameters_paired.sql
@@ -1,4 +1,0 @@
-ALTER TABLE SequencingParameters ADD COLUMN readLength2 int;
-UPDATE SequencingParameters SET readLength2 = paired * readLength;
-ALTER TABLE SequencingParameters MODIFY readLength2 int NOT NULL;
-ALTER TABLE SequencingParameters DROP COLUMN paired;